% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{LinkIt}
\alias{LinkIt}
\title{LinkIt}
\usage{
LinkIt(x,y,by,...)
}
\arguments{
\item{x, y}{data frames to be merged}

\item{by, by.x, by.y}{specifications of the columns used for merging.}

\item{control}{A list specifying how to process the alias text. See
``Details''.}
}
\value{
z The merged data frame.
}
\description{
Record linkage description.
}
\details{
LinkIt can automatically process the alias text for each dataset. In
`Control', users can specify the following options.

Set 'RemoveCommonWords' to TRUE to remove common words (those appearing in >
10\% of aliases).

Set `NormalizeSpaces' to TRUE to remove hanging whitespaces.

Set `RemovePunctuation' to TRUE to remove punctuation.

Set `ToLower' to TRUE to ignore case.
}
